requirements.txt
amqp==5.1.1
anyio==3.5.0
asgiref==3.5.0
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.6
celery-redbeat==2.0.0
click==8.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
Deprecated==1.2.13
fastapi==0.75.1
flower==1.0.0
greenlet==1.1.2
h11==0.13.0
humanize==4.0.0
idna==3.3
kombu==5.2.4
packaging==21.3
psycopg2-binary
prometheus-client==0.14.1
prompt-toolkit==3.0.29
pydantic==1.9.0
pyparsing==3.0.8
python-dateutil==2.8.2
pytz==2022.1
redis==4.2.2
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.36
starlette==0.17.1
tenacity==8.0.1
tornado==6.1
typing_extensions==4.1.1
uvicorn==0.17.6
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.0

